Metodo random

+++ ComeYaAPI/Models/DTOs/ItemDTOs/ReadItemDTO.cs
@@ -15,8 +15,8 @@
         public string Name { get; set; } = null!;
 
         public string Restaurant { get; set; } = null!;
-
-        public int Combo {  get; set; }
+       
+        public ulong Combo {  get; set; }
         public string MarketingImg1 { get; set; }
         public string MarketingImg2 { get; set; }
